Tests: New - Make QUnit test extendable.
