+++ tests/qunit/editor/editor-frontend.js
